add ext to old event ipv_miti2

+++ redshift-scripts/setup_process/alter_fact_dac_2931.sql
@@ -1,5 +1,5 @@
 ALTER TABLE  "conformed"."fact_user_journey_event" 
-ADD  COLUMN  event_timestamp_ms integer; 
+ADD  COLUMN  event_timestamp_ms BIGINT; 
 
 ALTER TABLE  "conformed"."fact_user_journey_event" 
 ADD  COLUMN  event_timestamp_ms_formatted varchar(1000);
